lab 2

+++ labs/lab2/hadoop/TotalCountReducer.java
@@ -1,12 +1,16 @@
 import java.io.*;
+
+import javax.naming.Context;
+
 import org.apache.hadoop.io.*;
 import org.apache.hadoop.mapreduce.*;
 import org.apache.hadoop.mapreduce.Reducer.*;
+import org.w3c.dom.Text;
 
 // Takes date and count of 1 as input and outputs the date with the total count
 
 public class TotalCountReducer 
-  extends Reducer<Text, IntWritable, Text, DoubleWritable> {
+  extends Reducer<Text, IntWritable, Text, IntWritable> {
    
    @Override
    public void reduce(Iterable<Text> date, Iterable<IntWritable> count, Context context) 
@@ -15,6 +19,6 @@ public class TotalCountReducer
         for(IntWritable c: count){
           sum += c.get();          
         } 
-        context.write(date, new DoubleWritable(sum));
+        context.write(date, new IntWritable(sum));
     }
 }
